instrumentation-genai/opentelemetry-instrumentation-llamaindex/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "splunk-otel-instrumentation-llamaindex"
dynamic = ["version"]
description = "OpenTelemetry Official LlamaIndex instrumentation"
readme = "README.rst"
license = "Apache-2.0"
requires-python = ">=3.9"
authors = [
    { name = "OpenTelemetry Authors", email = "cncf-opentelemetry-contributors@lists.cncf.io" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "opentelemetry-api ~= 1.38.0.dev0",
    "opentelemetry-instrumentation ~= 0.59b0.dev0",
    "opentelemetry-semantic-conventions ~= 0.59b0.dev0",
    "splunk-otel-util-genai>=0.1.4",
]

[project.optional-dependencies]
instruments = ["llama-index-core >= 0.14.0"]
test = [
    "llama-index-core >= 0.14.0",
    "llama-index-llms-openai >= 0.6.0",
    "pytest >= 7.0.0",
]

[project.entry-points.opentelemetry_instrumentor]
llamaindex = "opentelemetry.instrumentation.llamaindex:LlamaindexInstrumentor"

[project.urls]
Homepage = "https://github.com/open-telemetry/opentelemetry-python-contrib/tree/main/instrumentation-genai/opentelemetry-instrumentation-llamaindex"
Repository = "https://github.com/open-telemetry/opentelemetry-python-contrib"

[tool.hatch.version]
path = "src/opentelemetry/instrumentation/llamaindex/version.py"

[tool.hatch.build.targets.sdist]
include = ["/src", "/tests"]

[tool.hatch.build.targets.wheel]
packages = ["src/opentelemetry"]

[tool.ruff]
exclude = ["./"]
